Add likelihood wrapper class
