+++ gammapy/utils/fitting/__init__.py
@@ -1,6 +1,10 @@
 """Fitting framework and backends."""
+# Core
 from .parameter import *
 from .model import *
+from .likelihood import *
 from .fit import *
+
+# Backends
 from .iminuit import *
 from .sherpa import *
